requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
google-cloud-storage>=2.10.0
scipy>=1.11.0
matplotlib>=3.7.0
plotly>=5.17.0
seaborn>=0.12.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0
feedparser>=6.0.10
yfinance>=0.2.28
